Added a mock worker function
We can use this in unit tests. Instead of calling out to the actual runtime (which throws errors reading vm.SourceTextModule, something  complicated with the --experimental-vm-modules flag not getting passed to the ava thread), we create a worker which calls our simple mock function. All the worker lifecycle stuff is abstracted into a helper function which is used equally by the actual and mock workers - which gives us a really realistic mokc simulation.