describe-package: add rollup build
The current esbuild for this project seems to cause problems for other packages in this workspace. For example, compiler tests fail when importing anything from describe-package. I'm not sure what's going on here but using a rollup build, consistent with other packages, seems to help.

I've left the esbuild stuff in place.