force merge by default
In the CLI, we only force merge when the user asks us to. But here in project, we prefer to force merge. If you ask to merge two projects, we just merge them. Why woulldn't we? Add safeguads in the application, not down in the API

If it feels like a strange default its because the CLI should be doing all the force stuff, not the Project, but there we go